requirements.txt

inky
pillow
pyyaml
noass_sdk # if using weather
requests
json
textwrap
stockquotes # if using stock
bs4 # if using stock
lxml # if using stock